#154 Dashboard component modification started
